refactor: Separate story GET API call in GetStoryUseCase
