+++ frontend/Moment/app/src/main/java/snu/swpp/moment/ui/main_writeview/DailyViewModel.java
@@ -4,30 +4,24 @@ import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModel;
 import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
 import snu.swpp.moment.data.callback.MomentGetCallBack;
-import snu.swpp.moment.data.callback.StoryGetCallBack;
 import snu.swpp.moment.data.callback.TokenCallBack;
-import snu.swpp.moment.data.model.HashtagModel;
 import snu.swpp.moment.data.model.MomentPairModel;
-import snu.swpp.moment.data.model.StoryModel;
 import snu.swpp.moment.data.repository.AuthenticationRepository;
 import snu.swpp.moment.data.repository.MomentRepository;
 import snu.swpp.moment.data.repository.StoryRepository;
-import snu.swpp.moment.ui.main_writeview.uistate.StoryUiState;
 import snu.swpp.moment.ui.main_writeview.uistate.MomentUiState;
-import snu.swpp.moment.utils.EmotionMap;
+import snu.swpp.moment.ui.main_writeview.uistate.StoryUiState;
 import snu.swpp.moment.utils.TimeConverter;
 
 public class DailyViewModel extends ViewModel {
 
-    private final int REFRESH_TOKEN_EXPIRED = 1;
+    private final GetStoryUseCase getStoryUseCase;
     private final MutableLiveData<MomentUiState> momentState = new MutableLiveData<>();
-    private final MutableLiveData<StoryUiState> storyState = new MutableLiveData<>();
     private final AuthenticationRepository authenticationRepository;
     private final MomentRepository momentRepository;
-    private final StoryRepository storyRepository;
+
+    private final int REFRESH_TOKEN_EXPIRED = 1;
 
     public DailyViewModel(
         AuthenticationRepository authenticationRepository,
@@ -36,13 +30,13 @@ public class DailyViewModel extends ViewModel {
     ) {
         this.authenticationRepository = authenticationRepository;
         this.momentRepository = momentRepository;
-        this.storyRepository = storyRepository;
+        this.getStoryUseCase = new GetStoryUseCase(authenticationRepository, storyRepository);
     }
 
     public void getMoment(int year, int month, int date) {
         long[] dayInterval = TimeConverter.getOneDayIntervalTimestamps(year, month, date);
 
-        authenticationRepository.isTokenValid(new WriteViewTokenCallback() {
+        authenticationRepository.isTokenValid(new TokenCallBack() {
             @Override
             public void onSuccess() {
                 String access_token = authenticationRepository.getToken().getAccessToken();
@@ -63,50 +57,16 @@ public class DailyViewModel extends ViewModel {
                         }
                     });
             }
+
+            @Override
+            public void onFailure() {
+                momentState.setValue(new MomentUiState(REFRESH_TOKEN_EXPIRED, null));
+            }
         });
     }
 
     public void getStory(int year, int month, int date) {
-        long[] dayInterval = TimeConverter.getOneDayIntervalTimestamps(year, month, date);
-
-        authenticationRepository.isTokenValid(new WriteViewTokenCallback() {
-            @Override
-            public void onSuccess() {
-                String access_token = authenticationRepository.getToken().getAccessToken();
-                storyRepository.getStory(access_token, dayInterval[0], dayInterval[1],
-                    new StoryGetCallBack() {
-                        @Override
-                        public void onSuccess(ArrayList<StoryModel> story) {
-                            if (story.isEmpty()) {
-                                storyState.setValue(StoryUiState.empty());
-                            } else {
-                                StoryModel storyInstance = story.get(0);
-                                String title = storyInstance.getTitle();
-                                String content = storyInstance.getContent();
-                                int emotion = EmotionMap.getEmotionInt(storyInstance.getEmotion());
-                                List<HashtagModel> tags = storyInstance.getHashtags();
-                                List<String> parsedTags = new ArrayList<>();
-                                for (HashtagModel hashtagModel : tags) {
-                                    parsedTags.add(hashtagModel.getContent());
-                                }
-                                int score = storyInstance.getScore();
-                                Date createdAt = storyInstance.getCreatedAt();
-
-                                storyState.setValue(
-                                    new StoryUiState(null, false, title, content, emotion,
-                                        parsedTags,
-                                        score,
-                                        createdAt));
-                            }
-                        }
-
-                        @Override
-                        public void onFailure(Exception error) {
-                            storyState.setValue(StoryUiState.empty());
-                        }
-                    });
-            }
-        });
+        getStoryUseCase.getStory(year, month, date);
     }
 
     public void observeMomentState(Observer<MomentUiState> observer) {
@@ -114,14 +74,6 @@ public class DailyViewModel extends ViewModel {
     }
 
     public void observeStoryState(Observer<StoryUiState> observer) {
-        storyState.observeForever(observer);
-    }
-
-    abstract class WriteViewTokenCallback implements TokenCallBack {
-
-        @Override
-        public void onFailure() {
-            momentState.setValue(new MomentUiState(REFRESH_TOKEN_EXPIRED, null));
-        }
+        getStoryUseCase.observeStoryState(observer);
     }
 }
